# src/user/user.entity.ts
import { Exclude } from 'class-transformer';

export class User {
  id: string;
  login: string;

  @Exclude()
  password: string;

  version: number;
  createdAt: number;
  updatedAt: number;

  constructor(user: Partial<User>) {
    Object.assign(this, user);
  }
}